# Implement SchemaVisitorJCodec