--- a/modules/aws/src/smithy4s/json/awsJson.scala
+++ b/modules/aws/src/smithy4s/json/awsJson.scala
@@ -17,4 +17,4 @@
 package smithy4s.aws.json
 
 private[aws] object awsJson
-    extends smithy4s.http.json.JsonCodecAPI(AwsSchematicJCodec) {}
+    extends smithy4s.http.json.JsonCodecAPI(AwsSchematicJCodec, None) {}