--- a/modules/json/src/smithy4s/http/json/JsonCodecAPI.scala
+++ b/modules/json/src/smithy4s/http/json/JsonCodecAPI.scala
@@ -27,14 +27,20 @@ import smithy4s.schema.SchemaVisitor
 
 abstract class JsonCodecAPI(
     schemaVisitorJCodec: SchemaVisitor[JCodec],
+    hintMask: Option[HintMask] = None,
     readerConfig: ReaderConfig = JsonCodecAPI.defaultReaderConfig,
     writerConfig: WriterConfig = WriterConfig
 ) extends CodecAPI {
 
   type Codec[A] = JCodec[A]
 
-  def compileCodec[A](schema: Schema[A]): JCodec[A] =
+  def compileCodec[A](schema0: Schema[A]): JCodec[A] = {
+    val schema =
+      hintMask
+        .map(mask => schema0.transformHintsLocally(mask.apply))
+        .getOrElse(schema0)
     schema.compile(schemaVisitorJCodec)
+  }
 
   def mediaType[A](codec: JCodec[A]): HttpMediaType.Type =
     HttpMediaType("application/json")