… README.md
… modules/json/test/src/smithy4s/http/json/DocumentPropertyTests.scala
… modules/json/test/src/smithy4s/http/json/SchematicJCodecPropertyTests.scala
… modules/json/test/src/smithy4s/http/json/SchematicJCodecTests.scala