Please review commit by commit, but be aware:
1. not all commit compile/run, I went very granular to help with review
1. I did a `git mv` in one of the last commit but I think git still sees the file as a new file / deleted file

important remarks:
1. question: `def set` use `def expecting: String = "list"` and use `vectors` in error message, should we update that?
1. note: union* method use the whole alternative vector as argument rather than first and rest
1. called to .hints instead of .instance.hints in some place to fix the build