@daddykotex 2022-06-13
For example, one commit's diff: https://github.com/disneystreaming/smithy4s/pull/270/commits/eda5d47024cafc8b21ab9a3bed35cc4e27feef04