@Baccata 2022-06-14
One thing that's missing and that doesn't seem to have been caught by tests is that the hints on fields and alternatives must be added to the schema before the recursive call to `apply`. 

In other words, if you have something like 

```
structure Foo {
   @range(min: 1) 
   bar: Integer
} 
```

the constraint is not checked for. 

In all frankness, I think it might be indicative of a design mistake on my end on the `Schema` construct ... having to remember this is a burden for implementors. I should probably revert that particular decision