@Baccata 2022-06-14
Also, would you mind running jsoniter's benchmarks (smithy4s only, before/after) to check nothing has drastically changed.