@Baccata 2022-06-14
@daddykotex sorry, I meant the benchmarks added by Andriy in jsoniter. See https://github.com/plokhotnyuk/jsoniter-scala/pull/903.