@daddykotex 2022-06-15
Before:

```
[info] Benchmark                                  (size)   Mode  Cnt         Score        Error  Units
[info] ADTReading.smithy4sJson                       N/A  thrpt    5   2398176.220 ± 108303.195  ops/s
[info] ADTWriting.smithy4sJson                       N/A  thrpt    5   8437056.427 ±  85618.111  ops/s
[info] AnyValsReading.smithy4sJson                   N/A  thrpt    5   1959856.080 ±  29909.431  ops/s
[info] AnyValsWriting.smithy4sJson                   N/A  thrpt    5   8294473.824 ±  78167.130  ops/s
[info] ArrayOfBigDecimalsReading.smithy4sJson        128  thrpt    5    118015.048 ±   3855.693  ops/s
[info] ArrayOfBigDecimalsWriting.smithy4sJson        128  thrpt    5     82763.207 ±   1157.593  ops/s
[info] ArrayOfBigIntsReading.smithy4sJson            128  thrpt    5    161741.955 ±   7885.542  ops/s
[info] ArrayOfBigIntsWriting.smithy4sJson            128  thrpt    5     84357.276 ±   2622.033  ops/s
[info] ArrayOfBooleansReading.smithy4sJson           128  thrpt    5    303122.174 ±   2932.675  ops/s
[info] ArrayOfBooleansWriting.smithy4sJson           128  thrpt    5   1888649.419 ±  72666.724  ops/s
[info] ArrayOfBytesReading.smithy4sJson              128  thrpt    5    302592.700 ±   2923.858  ops/s
[info] ArrayOfBytesWriting.smithy4sJson              128  thrpt    5   1794201.254 ±   8833.376  ops/s
[info] ArrayOfDoublesReading.smithy4sJson            128  thrpt    5    204587.240 ±   2703.395  ops/s
[info] ArrayOfDoublesWriting.smithy4sJson            128  thrpt    5    216337.568 ±   3035.676  ops/s
[info] ArrayOfFloatsReading.smithy4sJson             128  thrpt    5    264770.635 ±   7574.607  ops/s
[info] ArrayOfFloatsWriting.smithy4sJson             128  thrpt    5    315041.260 ±   7335.917  ops/s
[info] ArrayOfInstantsReading.smithy4sJson           128  thrpt    5     12838.048 ±    452.895  ops/s
[info] ArrayOfInstantsWriting.smithy4sJson           128  thrpt    5     27186.557 ±    863.173  ops/s
[info] ArrayOfIntsReading.smithy4sJson               128  thrpt    5    263572.742 ±   2593.068  ops/s
[info] ArrayOfIntsWriting.smithy4sJson               128  thrpt    5    938403.109 ±  23404.162  ops/s
[info] ArrayOfLongsReading.smithy4sJson              128  thrpt    5    284957.780 ±   2817.141  ops/s
[info] ArrayOfLongsWriting.smithy4sJson              128  thrpt    5    588794.292 ±  33077.645  ops/s
[info] ArrayOfShortsReading.smithy4sJson             128  thrpt    5    283114.738 ±   3241.466  ops/s
[info] ArrayOfShortsWriting.smithy4sJson             128  thrpt    5   1405148.337 ±   3876.207  ops/s
[info] ArrayOfUUIDsReading.smithy4sJson              128  thrpt    5    136821.873 ±   1970.442  ops/s
[info] ArrayOfUUIDsWriting.smithy4sJson              128  thrpt    5    672070.895 ±   8401.650  ops/s
[info] Base64Reading.smithy4sJson                    128  thrpt    5   7875488.522 ± 101169.238  ops/s
[info] Base64Writing.smithy4sJson                    128  thrpt    5  10108998.119 ± 123852.620  ops/s
[info] BigDecimalWriting.smithy4sJson                128  thrpt    5    759872.675 ±  31113.731  ops/s
[info] BigIntReading.smithy4sJson                    128  thrpt    5   7779281.645 ± 123254.088  ops/s
[info] BigIntWriting.smithy4sJson                    128  thrpt    5    678839.809 ±  13162.155  ops/s
[info] ExtractFieldsReading.smithy4sJson             128  thrpt    5    152352.480 ±   2736.498  ops/s
[info] GeoJSONReading.smithy4sJson                   N/A  thrpt    5     23118.274 ±    173.196  ops/s
[info] GeoJSONWriting.smithy4sJson                   N/A  thrpt    5     20753.572 ±   1199.680  ops/s
[info] GitHubActionsAPIReading.smithy4sJson          N/A  thrpt    5    172020.411 ±   1390.046  ops/s
[info] GitHubActionsAPIWriting.smithy4sJson          N/A  thrpt    5    474886.344 ±  18327.983  ops/s
[info] GoogleMapsAPIPrettyPrinting.smithy4sJson      N/A  thrpt    5     41090.444 ±    463.633  ops/s
[info] GoogleMapsAPIReading.smithy4sJson             N/A  thrpt    5     16884.089 ±    260.824  ops/s
[info] GoogleMapsAPIWriting.smithy4sJson             N/A  thrpt    5     50990.538 ±    468.921  ops/s
[info] IntReading.smithy4sJson                       N/A  thrpt    5  57238862.853 ± 738785.297  ops/s
[info] IntWriting.smithy4sJson                       N/A  thrpt    5  48923954.353 ± 797891.915  ops/s
[info] ListOfBooleansReading.smithy4sJson            128  thrpt    5    354608.040 ±   5260.778  ops/s
[info] ListOfBooleansWriting.smithy4sJson            128  thrpt    5   2632389.648 ±  36261.687  ops/s
[info] MapOfIntsToBooleansReading.smithy4sJson       128  thrpt    5    120646.992 ±    735.708  ops/s
[info] MapOfIntsToBooleansWriting.smithy4sJson       128  thrpt    5    652794.469 ±  12557.574  ops/s
[info] MissingRequiredFieldsReading.smithy4sJson     N/A  thrpt    5   5281958.439 ± 115578.785  ops/s
[info] NestedStructsReading.smithy4sJson             128  thrpt    5     96460.334 ±   1944.129  ops/s
[info] NestedStructsWriting.smithy4sJson             128  thrpt    5    380156.696 ±   4613.754  ops/s
[info] OpenRTBReading.smithy4sJson                   N/A  thrpt    5    187514.571 ±   1350.416  ops/s
[info] OpenRTBWriting.smithy4sJson                   N/A  thrpt    5    353662.159 ±   2329.853  ops/s
[info] PrimitivesReading.smithy4sJson                N/A  thrpt    5   2007980.443 ±  21091.474  ops/s
[info] PrimitivesWriting.smithy4sJson                N/A  thrpt    5   8631353.129 ±  25746.832  ops/s
[info] SetOfIntsReading.smithy4sJson                 128  thrpt    5    115699.910 ±   2709.233  ops/s
[info] SetOfIntsWriting.smithy4sJson                 128  thrpt    5    984885.533 ±   4903.178  ops/s
[info] StringOfAsciiCharsReading.smithy4sJson        128  thrpt    5  15289783.977 ± 134781.651  ops/s
[info] StringOfAsciiCharsWriting.smithy4sJson        128  thrpt    5   6275547.578 ±  57086.814  ops/s
[info] StringOfEscapedCharsReading.smithy4sJson      128  thrpt    5   3415786.949 ± 148413.977  ops/s
[info] StringOfEscapedCharsWriting.smithy4sJson      128  thrpt    5   4141860.014 ±  15138.993  ops/s
[info] StringOfNonAsciiCharsReading.smithy4sJson     128  thrpt    5   3579543.017 ±  21416.760  ops/s
[info] StringOfNonAsciiCharsWriting.smithy4sJson     128  thrpt    5   3132251.030 ±  22949.191  ops/s
[info] TwitterAPIReading.smithy4sJson                N/A  thrpt    5     36879.833 ±    403.007  ops/s
[info] TwitterAPIWriting.smithy4sJson                N/A  thrpt    5     98619.046 ±   3614.063  ops/s
[info] VectorOfBooleansReading.smithy4sJson          128  thrpt    5    302640.060 ±   3472.153  ops/s
[info] VectorOfBooleansWriting.smithy4sJson          128  thrpt    5   1242409.330 ±   8016.772  ops/s
```


After:

```

[info] Benchmark                                  (size)   Mode  Cnt         Score         Error  Units
[info] ADTReading.smithy4sJson                       N/A  thrpt    5   2192986.055 ±   69796.055  ops/s
[info] ADTWriting.smithy4sJson                       N/A  thrpt    5   9147843.442 ±   89417.074  ops/s
[info] AnyValsReading.smithy4sJson                   N/A  thrpt    5   1523794.445 ±   18072.840  ops/s
[info] AnyValsWriting.smithy4sJson                   N/A  thrpt    5   8216562.967 ±  670717.674  ops/s
[info] ArrayOfBigDecimalsReading.smithy4sJson        128  thrpt    5    124568.582 ±    1130.280  ops/s
[info] ArrayOfBigDecimalsWriting.smithy4sJson        128  thrpt    5     85920.801 ±     313.407  ops/s
[info] ArrayOfBigIntsReading.smithy4sJson            128  thrpt    5    170648.766 ±     438.840  ops/s
[info] ArrayOfBigIntsWriting.smithy4sJson            128  thrpt    5     79859.746 ±    1975.276  ops/s
[info] ArrayOfBooleansReading.smithy4sJson           128  thrpt    5    317256.497 ±    3141.909  ops/s
[info] ArrayOfBooleansWriting.smithy4sJson           128  thrpt    5   1919922.819 ±   10379.568  ops/s
[info] ArrayOfBytesReading.smithy4sJson              128  thrpt    5    332181.792 ±    3005.566  ops/s
[info] ArrayOfBytesWriting.smithy4sJson              128  thrpt    5   1753380.343 ±  144820.477  ops/s
[info] ArrayOfDoublesReading.smithy4sJson            128  thrpt    5    222812.949 ±    3506.117  ops/s
[info] ArrayOfDoublesWriting.smithy4sJson            128  thrpt    5    208283.957 ±    2823.135  ops/s
[info] ArrayOfFloatsReading.smithy4sJson             128  thrpt    5    271796.217 ±    4980.277  ops/s
[info] ArrayOfFloatsWriting.smithy4sJson             128  thrpt    5    312308.873 ±    3774.360  ops/s
[info] ArrayOfInstantsReading.smithy4sJson           128  thrpt    5     12846.977 ±     328.487  ops/s
[info] ArrayOfInstantsWriting.smithy4sJson           128  thrpt    5     27691.117 ±     487.461  ops/s
[info] ArrayOfIntsReading.smithy4sJson               128  thrpt    5    313210.254 ±   17347.927  ops/s
[info] ArrayOfIntsWriting.smithy4sJson               128  thrpt    5    932004.126 ±   45867.784  ops/s
[info] ArrayOfLongsReading.smithy4sJson              128  thrpt    5    309755.100 ±    1990.987  ops/s
[info] ArrayOfLongsWriting.smithy4sJson              128  thrpt    5    598461.338 ±    3131.188  ops/s
[info] ArrayOfShortsReading.smithy4sJson             128  thrpt    5    327752.885 ±    1983.215  ops/s
[info] ArrayOfShortsWriting.smithy4sJson             128  thrpt    5   1282167.837 ±   54533.039  ops/s
[info] ArrayOfUUIDsReading.smithy4sJson              128  thrpt    5    195649.477 ±    3457.244  ops/s
[info] ArrayOfUUIDsWriting.smithy4sJson              128  thrpt    5    671052.970 ±    4077.451  ops/s
[info] Base64Reading.smithy4sJson                    128  thrpt    5   7547956.232 ± 1280662.235  ops/s
[info] Base64Writing.smithy4sJson                    128  thrpt    5   9765175.058 ±  670678.193  ops/s
[info] BigDecimalWriting.smithy4sJson                128  thrpt    5    707386.796 ±   15066.719  ops/s
[info] BigIntReading.smithy4sJson                    128  thrpt    5   7711269.635 ±  150767.700  ops/s
[info] BigIntWriting.smithy4sJson                    128  thrpt    5    685549.251 ±   14155.757  ops/s
[info] ExtractFieldsReading.smithy4sJson             128  thrpt    5    146412.366 ±    4449.215  ops/s
[info] GeoJSONReading.smithy4sJson                   N/A  thrpt    5     23217.532 ±     103.562  ops/s
[info] GeoJSONWriting.smithy4sJson                   N/A  thrpt    5     21628.545 ±     882.517  ops/s
[info] GitHubActionsAPIReading.smithy4sJson          N/A  thrpt    5    172817.299 ±    1398.592  ops/s
[info] GitHubActionsAPIWriting.smithy4sJson          N/A  thrpt    5    478474.476 ±   10031.970  ops/s
[info] GoogleMapsAPIPrettyPrinting.smithy4sJson      N/A  thrpt    5     38094.895 ±     529.542  ops/s
[info] GoogleMapsAPIReading.smithy4sJson             N/A  thrpt    5     16819.520 ±     145.982  ops/s
[info] GoogleMapsAPIWriting.smithy4sJson             N/A  thrpt    5     45923.026 ±    1942.469  ops/s
[info] IntReading.smithy4sJson                       N/A  thrpt    5  57237110.663 ±  528113.481  ops/s
[info] IntWriting.smithy4sJson                       N/A  thrpt    5  48545529.879 ± 1681527.083  ops/s
[info] ListOfBooleansReading.smithy4sJson            128  thrpt    5    356264.198 ±    4555.405  ops/s
[info] ListOfBooleansWriting.smithy4sJson            128  thrpt    5   2633806.317 ±   42024.375  ops/s
[info] MapOfIntsToBooleansReading.smithy4sJson       128  thrpt    5    124155.756 ±    1050.857  ops/s
[info] MapOfIntsToBooleansWriting.smithy4sJson       128  thrpt    5    683354.755 ±   22425.528  ops/s
[info] MissingRequiredFieldsReading.smithy4sJson     N/A  thrpt    5   5004746.466 ±  228617.676  ops/s
[info] NestedStructsReading.smithy4sJson             128  thrpt    5     95440.635 ±    1025.517  ops/s
[info] NestedStructsWriting.smithy4sJson             128  thrpt    5    374747.722 ±   41716.825  ops/s
[info] OpenRTBReading.smithy4sJson                   N/A  thrpt    5    188289.829 ±    3796.082  ops/s
[info] OpenRTBWriting.smithy4sJson                   N/A  thrpt    5    386514.174 ±    7588.826  ops/s
[info] PrimitivesReading.smithy4sJson                N/A  thrpt    5   1429732.700 ±   28112.858  ops/s
[info] PrimitivesWriting.smithy4sJson                N/A  thrpt    5   8308239.922 ±  408246.585  ops/s
[info] SetOfIntsReading.smithy4sJson                 128  thrpt    5    119789.079 ±    1937.745  ops/s
[info] SetOfIntsWriting.smithy4sJson                 128  thrpt    5    986635.158 ±    3518.802  ops/s
[info] StringOfAsciiCharsReading.smithy4sJson        128  thrpt    5  15049878.638 ±  738018.254  ops/s
[info] StringOfAsciiCharsWriting.smithy4sJson        128  thrpt    5   6291572.281 ±   24281.618  ops/s
[info] StringOfEscapedCharsReading.smithy4sJson      128  thrpt    5   3487470.924 ±  190335.583  ops/s
[info] StringOfEscapedCharsWriting.smithy4sJson      128  thrpt    5   4119490.415 ±   98874.969  ops/s
[info] StringOfNonAsciiCharsReading.smithy4sJson     128  thrpt    5   3563744.033 ±   25424.924  ops/s
[info] StringOfNonAsciiCharsWriting.smithy4sJson     128  thrpt    5   4476121.474 ±   94780.287  ops/s
[info] TwitterAPIReading.smithy4sJson                N/A  thrpt    5     34715.016 ±     833.851  ops/s
[info] TwitterAPIWriting.smithy4sJson                N/A  thrpt    5    103308.388 ±    4225.569  ops/s
[info] VectorOfBooleansReading.smithy4sJson          128  thrpt    5    303936.689 ±    1905.530  ops/s
[info] VectorOfBooleansWriting.smithy4sJson          128  thrpt    5   1660331.555 ±    5172.427  ops/s
```